requirements.txt
fastapi
uvicorn
python-docx
jinja2
groq
python-multipart
